// src/common/enums/response-code.enum.ts
export enum ResponseCode {
  // General
  SUCCESS = 'SUCCESS',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  INVALID_REQUEST = 'INVALID_REQUEST',

  // Authentication
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',

  // Validation
  VALIDATION_ERROR = 'VALIDATION_ERROR',

  // User
  USER_CREATED = 'USER.CREATED',
  USER_FOUND = 'USER.FOUND',
  USER_UPDATED = 'USER.UPDATED',
  USER_DELETED = 'USER.DELETED',
  USER_NOT_FOUND = 'USER.NOT_FOUND',
  USER_ALREADY_EXISTS = 'USER.ALREADY_EXISTS',
}